requirements.txt
python-telegram-bot==20.6 
gspread==5.7.2
google-auth==2.22.0
python-telegram-bot[job-queue]
matplotlib
pandas